requirements.txt
PyOpenGL
python-mpv
PyQt5
PyQt6
PySide2
PySide6
